refactor: separated students grades grid into its own component
- added support to `icon_size` attr in `<.profile_icon_with_name>`
